+++ doc/whats-new.rst
@@ -34,6 +34,8 @@ Deprecations
 Bug fixes
 ~~~~~~~~~
 
+- :py:attr:`DataArray.nbytes` now uses the ``nbytes`` property of the underlying array if available.
+  By `Max Jones <https://github.com/maxrjones>`_.
 
 Documentation
 ~~~~~~~~~~~~~